Spiiiiike

+++ db/record.go
@@ -28,29 +28,35 @@ func (r Record) String() string {
 
 func (r Record) Save(dir string) error {
 	fileName := fmt.Sprintf("%s/%s.pissy", dir, r.Uuid)
-	var buf bytes.Buffer
-	enc := gob.NewEncoder(&buf)
-	if err := enc.Encode(r); err != nil {
+	if exists, err := exists(fileName); err != nil {
 		return err
+	} else if !exists || r.Dirty {
+		var buf bytes.Buffer
+		enc := gob.NewEncoder(&buf)
+		if err := enc.Encode(r); err != nil {
+			return err
+		}
+		return ioutil.WriteFile(fileName, buf.Bytes(), 0600)
 	}
-	return ioutil.WriteFile(fileName, buf.Bytes(), 0600)
+	return nil
 }
 
-func (record *Record) Load(dir string, fileName string) error {
+func (r *Record) Load(dir string, fileName string) error {
 	buf, err := loadFile(dir, fmt.Sprintf("%s.pissy", fileName))
 	if err != nil {
 		return err
 	}
 	dec := gob.NewDecoder(buf)
-	return dec.Decode(record)
+	return dec.Decode(r)
 }
 
 func NewRecord() Record {
-	var record Record
 	now := time.Now()
-	record.Uuid = uuid.NewV4()
-	record.CreatedAt = now
-	record.UpdatedAt = now
-	record.Dirty = true
-	return record
+	uuid := uuid.NewV4()
+	return Record{
+		Uuid:      uuid,
+		CreatedAt: now,
+		UpdatedAt: now,
+		Dirty:     true,
+	}
 }
